# backend/src/main/resources/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "VieGym API",
    "description": "Backend API for VieGym workout, nutrition and AI coach",
    "contact": {
      "name": "VieGym"
    },
    "version": "v1"
  },
  "servers": [
    {
      "url": "http://localhost:8080/api/v1",
      "description": "Local development API v1"
    }
  ],
  "paths": {
...
    "/api/v1/workout-programs": {
      "get": {
        "tags": ["workout-program-controller"],
        "operationId": "listPrograms",
        "parameters": [
          { "name": "status", "in": "query", "required": false, "schema": { "type": "string" } },
          { "name": "page", "in": "query", "required": false, "schema": { "type": "integer", "default": 0 } },
          { "name": "size", "in": "query", "required": false, "schema": { "type": "integer", "default": 20 } }
        ],
        "responses": {
          "200": { "description": "OK", "content": { "*/*": { "schema": { "$ref": "#/components/schemas/ApiResponsePageResponseWorkoutProgramSummaryResponse" } } } }
        }
      },
      "post": {
        "tags": ["workout-program-controller"],
        "operationId": "createProgram",
        "requestBody": {
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/WorkoutProgramRequest" } } },
          "required": true
        },
        "responses": {
          "201": { "description": "Created", "content": { "*/*": { "schema": { "$ref": "#/components/schemas/ApiResponseWorkoutProgramResponse" } } } }
        }
      }
    },
    "/api/v1/workout-programs/{id}": {
      "get": {
        "tags": ["workout-program-controller"],
        "operationId": "getProgram",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer", "format": "int64" } }],
        "responses": {
          "200": { "description": "OK", "content": { "*/*": { "schema": { "$ref": "#/components/schemas/ApiResponseWorkoutProgramResponse" } } } }
        }
      },
      "put": {
        "tags": ["workout-program-controller"],
        "operationId": "updateProgram",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer", "format": "int64" } }],
        "requestBody": {
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/WorkoutProgramRequest" } } },
          "required": true
        },
        "responses": {
          "200": { "description": "OK", "content": { "*/*": { "schema": { "$ref": "#/components/schemas/ApiResponseWorkoutProgramResponse" } } } }
        }
      },
      "delete": {
        "tags": ["workout-program-controller"],
        "operationId": "archiveProgram",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer", "format": "int64" } }],
        "responses": {
          "200": { "description": "OK", "content": { "*/*": { "schema": { "$ref": "#/components/schemas/ApiResponseVoid" } } } }
        }
      }
    },
    "/api/v1/workout-schedules": {
      "get": {
        "tags": ["workout-schedule-controller"],
        "operationId": "listSchedules",
        "parameters": [
          { "name": "from", "in": "query", "required": false, "schema": { "type": "string", "format": "date" } },
          { "name": "to", "in": "query", "required": false, "schema": { "type": "string", "format": "date" } },
          { "name": "status", "in": "query", "required": false, "schema": { "type": "string" } },
          { "name": "page", "in": "query", "required": false, "schema": { "type": "integer", "default": 0 } },
          { "name": "size", "in": "query", "required": false, "schema": { "type": "integer", "default": 20 } }
        ],
        "responses": {
          "200": { "description": "OK", "content": { "*/*": { "schema": { "$ref": "#/components/schemas/ApiResponsePageResponseWorkoutScheduleResponse" } } } }
        }
      },
      "post": {
        "tags": ["workout-schedule-controller"],
        "operationId": "createSchedule",
        "requestBody": {
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/WorkoutScheduleRequest" } } },
          "required": true
        },
        "responses": {
          "201": { "description": "Created", "content": { "*/*": { "schema": { "$ref": "#/components/schemas/ApiResponseWorkoutScheduleResponse" } } } }
        }
      }
    },
    "/api/v1/workout-schedules/{id}": {
      "get": {
        "tags": ["workout-schedule-controller"],
        "operationId": "getSchedule",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer", "format": "int64" } }],
        "responses": {
          "200": { "description": "OK", "content": { "*/*": { "schema": { "$ref": "#/components/schemas/ApiResponseWorkoutScheduleResponse" } } } }
        }
      },
      "put": {
        "tags": ["workout-schedule-controller"],
        "operationId": "updateSchedule",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer", "format": "int64" } }],
        "requestBody": {
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/WorkoutScheduleRequest" } } },
          "required": true
        },
        "responses": {
          "200": { "description": "OK", "content": { "*/*": { "schema": { "$ref": "#/components/schemas/ApiResponseWorkoutScheduleResponse" } } } }
        }
      },
      "delete": {
        "tags": ["workout-schedule-controller"],
        "operationId": "cancelSchedule",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer", "format": "int64" } },
          { "name": "cancelReason", "in": "query", "required": false, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "OK", "content": { "*/*": { "schema": { "$ref": "#/components/schemas/ApiResponseVoid" } } } }
        }
      }
    },
    "/api/v1/workout-sessions/{scheduleId}/start": {
      "post": {
        "tags": ["workout-session-controller"],
        "operationId": "startSession",
        "parameters": [{ "name": "scheduleId", "in": "path", "required": true, "schema": { "type": "integer", "format": "int64" } }],
        "responses": {
          "201": { "description": "Created", "content": { "*/*": { "schema": { "$ref": "#/components/schemas/ApiResponseStartSessionResponse" } } } }
        }
      }
    },
    "/api/v1/workout-sessions/{id}/pause": {
      "post": {
        "tags": ["workout-session-controller"],
        "operationId": "pauseSession",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer", "format": "int64" } }],
        "responses": {
          "200": { "description": "OK", "content": { "*/*": { "schema": { "$ref": "#/components/schemas/ApiResponsePauseSessionResponse" } } } }
        }
      }
    },
    "/api/v1/workout-sessions/{id}/resume": {
      "post": {
        "tags": ["workout-session-controller"],
        "operationId": "resumeSession",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer", "format": "int64" } }],
        "responses": {
          "200": { "description": "OK", "content": { "*/*": { "schema": { "$ref": "#/components/schemas/ApiResponseResumeSessionResponse" } } } }
        }
      }
    },
    "/api/v1/workout-sessions/{id}/finish": {
      "post": {
        "tags": ["workout-session-controller"],
        "operationId": "finishSession",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer", "format": "int64" } }],
        "requestBody": {
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/FinishSessionRequest" } } },
          "required": true
        },
        "responses": {
          "200": { "description": "OK", "content": { "*/*": { "schema": { "$ref": "#/components/schemas/ApiResponseFinishSessionResponse" } } } }
        }
      }
    },
    "/api/v1/workout-sessions/{id}/discard": {
      "post": {
        "tags": ["workout-session-controller"],
        "operationId": "discardSession",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer", "format": "int64" } }],
        "requestBody": {
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/DiscardSessionRequest" } } },
          "required": false
        },
        "responses": {
          "200": { "description": "OK", "content": { "*/*": { "schema": { "$ref": "#/components/schemas/ApiResponseVoid" } } } }
        }
      }
    },
    "/api/v1/workout-logs": {
      "get": {
        "tags": ["workout-log-controller"],
        "operationId": "listWorkoutLogs",
        "parameters": [
          { "name": "from", "in": "query", "required": false, "schema": { "type": "string", "format": "date" } },
          { "name": "to", "in": "query", "required": false, "schema": { "type": "string", "format": "date" } },
          { "name": "exerciseId", "in": "query", "required": false, "schema": { "type": "integer", "format": "int64" } },
          { "name": "page", "in": "query", "required": false, "schema": { "type": "integer", "default": 0 } },
          { "name": "size", "in": "query", "required": false, "schema": { "type": "integer", "default": 20 } }
        ],
        "responses": {
          "200": { "description": "OK", "content": { "*/*": { "schema": { "$ref": "#/components/schemas/ApiResponsePageResponseWorkoutLogSummaryDto" } } } }
        }
      }
    },
    "/api/v1/workout-logs/{id}": {
      "get": {
        "tags": ["workout-log-controller"],
        "operationId": "getWorkoutLogDetail",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer", "format": "int64" } }],
        "responses": {
          "200": { "description": "OK", "content": { "*/*": { "schema": { "$ref": "#/components/schemas/ApiResponseWorkoutLogDetailDto" } } } }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApiResponseVoid": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          },
          "data": {}
        }
      },
...
      "WorkoutExerciseRequest": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "format": "int64" },
          "exerciseId": { "type": "integer", "format": "int64" },
          "sortOrder": { "type": "integer" },
          "targetSets": { "type": "integer" },
          "targetRepsMin": { "type": "integer" },
          "targetRepsMax": { "type": "integer" },
          "targetDurationSeconds": { "type": "integer" },
          "restSeconds": { "type": "integer" },
          "note": { "type": "string" }
        }
      },
      "WorkoutDayRequest": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "format": "int64" },
          "dayNumber": { "type": "integer" },
          "name": { "type": "string" },
          "note": { "type": "string" },
          "exercises": { "type": "array", "items": { "$ref": "#/components/schemas/WorkoutExerciseRequest" } }
        }
      },
      "WorkoutProgramRequest": {
        "type": "object",
        "properties": {
          "name": { "type": "string" },
          "programType": { "type": "string" },
          "description": { "type": "string" },
          "status": { "type": "string" },
          "days": { "type": "array", "items": { "$ref": "#/components/schemas/WorkoutDayRequest" } }
        }
      },
...
      "WorkoutDayResponse": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "format": "int64" },
          "dayNumber": { "type": "integer" },
          "name": { "type": "string" },
          "note": { "type": "string" },
          "exercises": { "type": "array", "items": { "$ref": "#/components/schemas/WorkoutExerciseResponse" } }
        }
      },
      "WorkoutProgramResponse": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "format": "int64" },
          "name": { "type": "string" },
          "programType": { "type": "string" },
          "description": { "type": "string" },
          "status": { "type": "string" },
          "days": { "type": "array", "items": { "$ref": "#/components/schemas/WorkoutDayResponse" } }
        }
      },
      "WorkoutProgramSummaryResponse": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "format": "int64" },
          "name": { "type": "string" },
          "programType": { "type": "string" },
          "description": { "type": "string" },
          "status": { "type": "string" },
          "totalDays": { "type": "integer" }
        }
      },
      "ApiResponseWorkoutProgramResponse": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean" },
          "data": { "$ref": "#/components/schemas/WorkoutProgramResponse" },
          "error": { "$ref": "#/components/schemas/ApiError" }
        }
      },
      "ApiResponsePageResponseWorkoutProgramSummaryResponse": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean" },
          "data": {
            "type": "object",
            "properties": {
              "content": { "type": "array", "items": { "$ref": "#/components/schemas/WorkoutProgramSummaryResponse" } },
              "page": { "type": "integer" },
              "size": { "type": "integer" },
              "totalElements": { "type": "integer", "format": "int64" },
              "totalPages": { "type": "integer" },
              "first": { "type": "boolean" },
              "last": { "type": "boolean" }
            }
          },
          "error": { "$ref": "#/components/schemas/ApiError" }
        }
      },
      "WorkoutScheduleRequest": {
        "type": "object",
        "properties": {
          "workoutProgramId": { "type": "integer", "format": "int64" },
          "workoutDayId": { "type": "integer", "format": "int64" },
          "scheduledDate": { "type": "string", "format": "date" },
          "scheduledTime": { "type": "string" },
          "title": { "type": "string" }
        }
      },
      "WorkoutScheduleResponse": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "format": "int64" },
          "workoutProgramId": { "type": "integer", "format": "int64" },
          "workoutDayId": { "type": "integer", "format": "int64" },
          "scheduledDate": { "type": "string", "format": "date" },
          "scheduledTime": { "type": "string" },
          "title": { "type": "string" },
          "status": { "type": "string" },
          "cancelReason": { "type": "string" },
          "completedAt": { "type": "string", "format": "date-time" },
          "activeSessionId": { "type": "integer", "format": "int64" }
        }
      },
      "ApiResponseWorkoutScheduleResponse": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean" },
          "data": { "$ref": "#/components/schemas/WorkoutScheduleResponse" },
          "error": { "$ref": "#/components/schemas/ApiError" }
        }
      },
      "ApiResponsePageResponseWorkoutScheduleResponse": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean" },
          "data": {
            "type": "object",
            "properties": {
              "content": { "type": "array", "items": { "$ref": "#/components/schemas/WorkoutScheduleResponse" } },
              "page": { "type": "integer" },
              "size": { "type": "integer" },
              "totalElements": { "type": "integer", "format": "int64" },
              "totalPages": { "type": "integer" },
              "first": { "type": "boolean" },
              "last": { "type": "boolean" }
            }
          },
          "error": { "$ref": "#/components/schemas/ApiError" }
        }
      },
      "StartSessionResponse": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "format": "int64" },
          "workoutScheduleId": { "type": "integer", "format": "int64" },
          "status": { "type": "string" },
          "startedAt": { "type": "string", "format": "date-time" },
          "totalPausedSeconds": { "type": "integer" },
          "exercises": { "type": "array", "items": { "$ref": "#/components/schemas/WorkoutExerciseResponse" } }
        }
      },
      "ApiResponseStartSessionResponse": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean" },
          "data": { "$ref": "#/components/schemas/StartSessionResponse" },
          "error": { "$ref": "#/components/schemas/ApiError" }
        }
      },
      "PauseSessionResponse": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "format": "int64" },
          "status": { "type": "string" },
          "pausedAt": { "type": "string", "format": "date-time" },
          "totalPausedSeconds": { "type": "integer" }
        }
      },
      "ApiResponsePauseSessionResponse": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean" },
          "data": { "$ref": "#/components/schemas/PauseSessionResponse" },
          "error": { "$ref": "#/components/schemas/ApiError" }
        }
      },
      "ResumeSessionResponse": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "format": "int64" },
          "status": { "type": "string" },
          "totalPausedSeconds": { "type": "integer" }
        }
      },
      "ApiResponseResumeSessionResponse": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean" },
          "data": { "$ref": "#/components/schemas/ResumeSessionResponse" },
          "error": { "$ref": "#/components/schemas/ApiError" }
        }
      },
      "FinishSetLogRequest": {
        "type": "object",
        "properties": {
          "setNumber": { "type": "integer" },
          "reps": { "type": "integer" },
          "weightKg": { "type": "number" },
          "durationSeconds": { "type": "integer" },
          "rpe": { "type": "number" },
          "completed": { "type": "boolean" }
        }
      },
      "FinishExerciseLogRequest": {
        "type": "object",
        "properties": {
          "exerciseId": { "type": "integer", "format": "int64" },
          "sortOrder": { "type": "integer" },
          "durationSeconds": { "type": "integer" },
          "completed": { "type": "boolean" },
          "sets": { "type": "array", "items": { "$ref": "#/components/schemas/FinishSetLogRequest" } }
        }
      },
      "FinishSessionRequest": {
        "type": "object",
        "properties": {
          "note": { "type": "string" },
          "exercises": { "type": "array", "items": { "$ref": "#/components/schemas/FinishExerciseLogRequest" } }
        }
      },
...
      "FinishSessionResponse": {
        "type": "object",
        "properties": {
          "workoutLogId": { "type": "integer", "format": "int64" },
          "workoutSessionId": { "type": "integer", "format": "int64" },
          "durationSeconds": { "type": "integer" },
          "totalVolumeKg": { "type": "number" },
          "newPersonalRecords": { "type": "array", "items": { "$ref": "#/components/schemas/PersonalRecordDto" } }
        }
      },
      "ApiResponseFinishSessionResponse": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean" },
          "data": { "$ref": "#/components/schemas/FinishSessionResponse" },
          "error": { "$ref": "#/components/schemas/ApiError" }
        }
      },
      "DiscardSessionRequest": {
        "type": "object",
        "properties": {
          "reason": { "type": "string" }
        }
      },
...
      "WorkoutLogSummaryDto": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "format": "int64" },
          "workoutSessionId": { "type": "integer", "format": "int64" },
          "workoutScheduleId": { "type": "integer", "format": "int64" },
          "title": { "type": "string" },
          "startedAt": { "type": "string", "format": "date-time" },
          "completedAt": { "type": "string", "format": "date-time" },
          "durationSeconds": { "type": "integer" },
          "totalVolumeKg": { "type": "number" },
          "totalExercises": { "type": "integer" },
          "prCount": { "type": "integer" }
        }
      },
      "WorkoutLogDetailDto": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "format": "int64" },
          "workoutSessionId": { "type": "integer", "format": "int64" },
          "workoutScheduleId": { "type": "integer", "format": "int64" },
          "title": { "type": "string" },
          "startedAt": { "type": "string", "format": "date-time" },
          "completedAt": { "type": "string", "format": "date-time" },
          "durationSeconds": { "type": "integer" },
          "totalVolumeKg": { "type": "number" },
          "note": { "type": "string" },
          "exercises": { "type": "array", "items": { "$ref": "#/components/schemas/WorkoutExerciseLogDto" } },
          "personalRecords": { "type": "array", "items": { "$ref": "#/components/schemas/PersonalRecordDto" } }
        }
      },
      "ApiResponseWorkoutLogDetailDto": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean" },
          "data": { "$ref": "#/components/schemas/WorkoutLogDetailDto" },
          "error": { "$ref": "#/components/schemas/ApiError" }
        }
      },
      "ApiResponsePageResponseWorkoutLogSummaryDto": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean" },
          "data": {
            "type": "object",
            "properties": {
              "content": { "type": "array", "items": { "$ref": "#/components/schemas/WorkoutLogSummaryDto" } },
              "page": { "type": "integer" },
              "size": { "type": "integer" },
              "totalElements": { "type": "integer", "format": "int64" },
              "totalPages": { "type": "integer" },
              "first": { "type": "boolean" },
              "last": { "type": "boolean" }
            }
          },
          "error": { "$ref": "#/components/schemas/ApiError" }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}